Add files via upload
In this code we make a function inside the class, which calculates the volume and prints it. We do not need to calculate the volume in the main function. We can just call the function from the class, using the " . " operator. Eg. cube1.volume, (name_of_variable).(name_of_function)