Add files via upload
In this code we make two functions inside the class.One to take input and the other to calculate volume and print it. We call both the functions in the main function. Therefore we do not need to take input or calculate any volume , we just have to call the functions.